Create P2P peer book utils
